requirements.txt
streamlit
pandas
numpy
requests
python-dotenv
openai
pinecone-client
